+++ Modules/Intent.Modules.Common/Templates/StaticContent/StaticBinaryContentTemplate.cs
@@ -49,7 +49,7 @@ namespace Intent.Modules.Common.Templates.StaticContent
         /// <inheritdoc />
         public override string GetCorrelationId()
         {
-            return $"{Id}#{_relativeOutputPath}";
+            return $"{Id}#{OutputTarget.Id}#{_relativeOutputPath}";
         }
     }
 }