+++ Tests/ArchTemplateResearch/ArchTemplateResearch.Api/Filters/ExceptionFilter.cs
@@ -1,0 +1,55 @@
+using System.Diagnostics;
+using ArchTemplateResearch.Application.Common.Exceptions;
+using ArchTemplateResearch.Domain.Common.Exceptions;
+using Intent.RoslynWeaver.Attributes;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Filters;
+
+[assembly: DefaultIntentManaged(Mode.Fully)]
+[assembly: IntentTemplate("Intent.AspNetCore.Controllers.ExceptionFilter", Version = "1.0")]
+
+namespace ArchTemplateResearch.Api.Filters
+{
+    public class ExceptionFilter : Microsoft.AspNetCore.Mvc.Filters.IExceptionFilter
+    {
+        public void OnException(ExceptionContext context)
+        {
+            switch (context.Exception)
+            {
+                case ForbiddenAccessException:
+                    context.Result = new ForbidResult();
+                    context.ExceptionHandled = true;
+                    break;
+                case UnauthorizedAccessException:
+                    context.Result = new UnauthorizedResult();
+                    context.ExceptionHandled = true;
+                    break;
+                case NotFoundException exception:
+                    context.Result = new NotFoundObjectResult(new ProblemDetails
+                    {
+                        Detail = exception.Message
+                    })
+                    .AddContextInformation(context);
+                    context.ExceptionHandled = true;
+                    break;
+                default:
+                    break;
+            }
+        }
+    }
+
+    internal static class ProblemDetailsExtensions
+    {
+        public static IActionResult AddContextInformation(this ObjectResult objectResult, ExceptionContext context)
+        {
+            if (objectResult.Value is not ProblemDetails problemDetails)
+            {
+                return objectResult;
+            }
+
+            problemDetails.Extensions.Add("traceId", Activity.Current?.Id ?? context.HttpContext.TraceIdentifier);
+            problemDetails.Type = "https://httpstatuses.io/" + (objectResult.StatusCode ?? problemDetails.Status);
+            return objectResult;
+        }
+    }
+}