+++ Modules/Intent.Modules.Modelers.CodebaseStructure/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.2
+
+- Fixed: Relocated the `Root Folder Options` stereotype to the `Codebase Structure` designer so as to control the root folder options for generated content.
+
 ### Version 1.0.1
 
 - Improvement: Moved common folder logic from `Intent.VisualStudio.Projects`